fix(playerbot): Fix GetPrimarySpecialization type conversion and syntax errors
Fixed C2440 type conversion and C2059 syntax errors:

1. Cast ChrSpecialization to uint8 where needed:
   - HealingNodes.h line 692: Added uint8() cast
   - MovementNodes.h line 498: Added uint8() cast
   - UtilityContextBuilder.h line 159: Added uint8() cast
   - HybridAIController.cpp line 94: Added uint8() cast

2. Fixed extra closing parenthesis:
   - UtilityContextBuilder.h line 159: Removed extra )
   - HybridAIController.cpp line 94: Removed extra )

TrinityCore changed GetPrimarySpecialization() return type from uint8 to ChrSpecialization enum.
Need explicit cast when assigning to uint8 variables.

Total: ~8 C2440 + C2059 errors fixed

+++ src/modules/Playerbot/AI/BehaviorTree/Nodes/MovementNodes.h
@@ -495,7 +495,7 @@ public:
 
             // Check if member is healer
             uint8 classId = member->GetClass();
-            uint8 spec = member->GetPrimarySpecialization();
+            uint8 spec = uint8(member->GetPrimarySpecialization());
 
             bool isHealer = false;
             if (classId == CLASS_PRIEST && (spec == 1 || spec == 2)) isHealer = true;
