+++ src/modules/Playerbot/AI/Utility/UtilityContextBuilder.h
@@ -156,7 +156,7 @@ private:
             return UtilityContext::Role::DPS;
 
         uint8 classId = bot->GetClass();
-        uint8 spec = bot->GetPrimarySpecialization());
+        uint8 spec = uint8(bot->GetPrimarySpecialization());
 
         // Tank specs
         if (classId == CLASS_WARRIOR && spec == 2) // Protection
